Moving section component to components/layout
